requirements.txt
scholarly==1.7.11
python-slugify==8.0.4
requests==2.32.3


